backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
